template converter/madewithloveuploadingassests!/requirements.txt
selenium
requests
Pillow
